@@ -1,4 +1,5 @@
-import React, { Component, PropTypes } from "react";
+import React, { Component } from "react";
+import PropTypes from "prop-types";
 import { connect } from "react-redux";
 import {
     task as TaskActions
@@ -11,7 +12,7 @@ import util from "../../utils";
 import resx from "../../resources";
 
 /*eslint-disable quotes*/
-const svgIcon = require(`!raw!./../svg/edit.svg`);
+const svgIcon = require(`!raw-loader!./../svg/edit.svg`);
 
 class TopPane extends Component {
     constructor() {
@@ -21,7 +22,7 @@ class TopPane extends Component {
         };
     }
 
-    componentWillMount() {
+    componentDidMount() {
         const {props} = this;
         const persistedSettings = util.utilities.persistent.load();
 
