@@ -10,6 +10,9 @@ const BundleAnalyzerPlugin = require('webpack-bundle-analyzer').BundleAnalyzerPl
 module.exports = {
     context: path.resolve(__dirname, '.'),
     entry: "./src/main.jsx",
+    optimization: {
+        minimize: isProduction
+    },
     output: {
         path: path.resolve(__dirname, '../admin/personaBar/scripts/bundles/'),
         publicPath: isProduction ? "" : "http://localhost:8100/dist/",
@@ -66,8 +69,6 @@ module.exports = {
     isProduction
         ?
         [
-            new webpack.optimize.UglifyJsPlugin(),
-            new webpack.optimize.DedupePlugin(),
             new webpack.DefinePlugin({
                 VERSION: JSON.stringify(packageJson.version),
                 "process.env": {
