@@ -1,6 +1,6 @@
-import React, { Component, PropTypes } from "react";
+import React, { Component } from "react";
+import PropTypes from "prop-types";
 import { connect } from "react-redux";
-import ReactDOM from "react-dom";
 import styles from "./style.less";
 import GridSystem from "dnn-grid-system";
 import LeftPane from "../LeftPane";
@@ -38,7 +38,7 @@ class TaskStatusItemRow extends Component {
         // the "findDOMNode was called on an unmounted component." error we need to check if the component is mounted before execute this code
         if (!this._isMounted) { return; }
 
-        if (!ReactDOM.findDOMNode(this).contains(event.target) && (typeof event.target.className == "string" && event.target.className.indexOf("do-not-close") == -1)) {
+        if (!this.node.contains(event.target) && (typeof event.target.className == "string" && event.target.className.indexOf("do-not-close") == -1)) {
 
             this.timeout = 475;
         } else {
@@ -50,7 +50,7 @@ class TaskStatusItemRow extends Component {
     render() {
         const {props} = this;
         return (
-            <div className={styles.taskStatusItemRow}>
+            <div className={styles.taskStatusItemRow} ref={node => this.node = node}>
                 <GridSystem>
                     <LeftPane
                         friendlyName={props.friendlyName}
@@ -60,7 +60,7 @@ class TaskStatusItemRow extends Component {
                         elapsedTime={props.elapsedTime}
                         startDate={props.startDate}
                         key={"schedule-left-" + props.scheduleId}
-                        />
+                    />
                     <RightPane
                         scheduleId={props.scheduleId}
                         servers={props.servers}
@@ -68,7 +68,7 @@ class TaskStatusItemRow extends Component {
                         threadId={props.threadId}
                         scheduleSource={props.scheduleSource}
                         key={"schedule-right-" + props.scheduleId}
-                        />
+                    />
                 </GridSystem>
             </div>
         );
