@@ -4,8 +4,8 @@ import LogMonitor from "redux-devtools-log-monitor";
 import DockMonitor from "redux-devtools-dock-monitor";
 
 export default createDevTools(
-  <DockMonitor toggleVisibilityKey="ctrl-h"
-               changePositionKey="ctrl-q">
-    <LogMonitor />
-  </DockMonitor>
+    <DockMonitor toggleVisibilityKey="ctrl-h"
+        changePositionKey="ctrl-q">
+        <LogMonitor />
+    </DockMonitor>
 );
